Change connComp to mask
Connected component 0 in a given frame is not necessarily the one with the most pixels. So when frames are stitched together, one will be dominantly connComp 1 and another connComp 0, leading to an apparent discontinuity. Instead, make connComp now a mask highlighting pixels which do not match the dominant connected component of that frame. That way, they will stitch fine and only highlight aberrant pixels.